requirements.txt
python-dotenv
jupyterlab
ipywidgets
requests
numpy
pandas
scipy
scikit-learn
matplotlib
gensim
torch
google-generativeai
transformers
tqdm
openai
gradio
datasets
sentencepiece
anthropic
google-generativeai
unstructured
chromadb
plotly
jupyter-dash
beautifulsoup4
modal
ollama
accelerate
bitsandbytes
psutil
setuptools
speedtest-cli
litellm